requirements.txt
# Core dependencies
python-dotenv==1.0.0
openai==1.76.0
anthropic==0.50.0
google-generativeai==0.3.2
lxml_html_clean==0.4.2

# Web tools
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
newspaper3k==0.2.8
serpapi==0.1.5

# Data processing
pandas==2.1.1
numpy==1.26.0
nltk==3.8.1

# Web application
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
responses==0.23.3